[orca-agent] integrate assertHttpController for swagger-based agent API
