+++ packages/cli/src/utils.ts
@@ -1,7 +1,7 @@
 export function getAgentDescription(agentName: string): string {
     const descriptions = {
       "orca": "AI agent for orchestration",
-      "orca-mcp-server": "Direct MCP server call agent"
+      "orca-api-server": "Direct API server call"
     } as { [key: string]: string };
     return descriptions[agentName] || "AI conversational agent";
   }
