+++ packages/orca-agent/package.json
@@ -20,11 +20,10 @@
   "dependencies": {
     "@agentica/core": "^0.27.3",
     "@agentica/rpc": "^0.27.3",
-    "@samchon/openapi": "^3.3.0",
     "dotenv": "^16.4.7",
     "openai": "^4.87.3",
+    "redis": "^5.5.6",
     "tgrid": "^1.1.0",
-    "typia": "^8.2.0",
-    "redis": "^5.5.6"
+    "typia": "^8.2.0"
   }
 }
