+++ packages/orca-api-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@orchestrica/mcp-server",
+  "name": "@orchestrica/api-server",
   "version": "0.0.1",
   "main": "dist/main.js",
   "scripts": {
