+++ packages/orca-api-server/src/routes/memory.ts
@@ -12,6 +12,13 @@ const router = Router();
  * /memory/session/{id}:
  *   get:
  *     summary: Get session memory state
+ *     description: |
+ *       Retrieve the current memory state of a session from Redis.
+ *       This memory stores historical context and variables used during previous LLM interactions.
+ *       It allows the system to maintain continuity across multiple user prompts.
+ *
+ *       @param id - The unique identifier of the session (e.g., UUID).
+ *       @tag memory
  *     parameters:
  *       - in: path
  *         name: id
@@ -29,6 +36,14 @@ router.get("/session/:id", getSessionState);
  * /memory/session/{id}:
  *   post:
  *     summary: Save session memory state
+ *     description: |
+ *       Save or update the session memory state in Redis for a specific session ID.
+ *       This memory may include context history, variables, or other LLM-relevant state to be used in future tasks.
+ *       The memory is stored in Redis and will be referenced in future LLM function calls for context continuity.
+ *
+ *       @param id - The session identifier, matching the key used for storage and retrieval.
+ *       @body - JSON object representing session memory to store.
+ *       @tag memory
  *     parameters:
  *       - in: path
  *         name: id
@@ -52,6 +67,11 @@ router.post("/session/:id", saveSessionState);
  * /memory/template/flush:
  *   post:
  *     summary: Flush in-memory template store
+ *     description: |
+ *       Flush (clear) the in-memory cache of prompt templates stored in Redis.
+ *       Use this endpoint when you need to reset or refresh the current template state, for example after updating a template.
+ *
+ *       @tag memory
  *     responses:
  *       200:
  *         description: Template cache flushed
